+++ X10D.Unity.Tests/Assets/Tests/TestBehaviour.cs
@@ -1,0 +1,10 @@
+﻿namespace X10D.Unity.Tests
+{
+    internal sealed class TestBehaviour : Singleton<TestBehaviour>
+    {
+        public bool Flag
+        {
+            get => true;
+        }
+    }
+}